+++ src/main/java/lotto/domain/Lotto.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
