+++ src/main/java/lotto/domain/LottoFactory.java
@@ -1,10 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class LottoFactory {
     private static final LottoNumber[] LOTTO_NUMBERS = {
