+++ src/main/java/lotto/view/SummaryPrinter.java
@@ -1,4 +1,9 @@
-package lotto;
+package lotto.view;
+
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.Lottos;
+import lotto.domain.Winners;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
